add split patching api
