+++ python/nano/src/bigdl/nano/pytorch/__init__.py
@@ -21,6 +21,7 @@ __all__ = ["Trainer", "TorchNano", "InferenceOptimizer"]
 import os
 if 'KMP_INIT_AT_FORK' in os.environ:
     del os.environ['KMP_INIT_AT_FORK']
+from .dispatcher import patch_torch, unpatch_torch
 from bigdl.nano.pytorch.inference import InferenceOptimizer
 from bigdl.nano.pytorch.trainer import Trainer
 from bigdl.nano.pytorch.torch_nano import TorchNano
