+++ python/nano/src/bigdl/nano/pytorch/dispatcher.py
@@ -1,0 +1,76 @@
+#
+# Copyright 2016 The BigDL Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+from importlib.util import find_spec
+
+
+_mapping_torch = None
+
+
+def _get_patch_map():
+    global _mapping_torch
+
+    # decide if generate 
+    patch_torch = find_spec("pytorch_lightning") is not None
+
+    if patch_torch and _mapping_torch is None:
+        _mapping_torch = []
+        import pytorch_lightning
+        import torchvision
+        from bigdl.nano.pytorch import Trainer
+        from bigdl.nano.pytorch.vision import transforms
+        from bigdl.nano.pytorch.vision import datasets
+        _mapping_torch += [
+            [pytorch_lightning, "Trainer", Trainer, None],
+            [torchvision, "transforms", transforms, None],
+            [torchvision, "datasets", datasets, None],
+        ]
+
+    if not patch_torch:
+        _mapping_torch = []
+
+    return _mapping_torch
+
+
+def patch_torch():
+    """
+    patch_torch is used to patch optimized torch classes to replace original ones.
+
+    Optimized classes include:
+
+    | 1. pytorch_lightning.Trainer -> bigdl.nano.pytorch.Trainer
+    | 2. torchvision.transforms -> bigdl.nano.pytorch.vision.transforms
+    | 3. torchvision.datasets -> bigdl.nano.pytorch.vision.datasets
+    """
+    # TODO: patch_torch to support gpu-to-cpu patching
+
+    mapping_torch = _get_patch_map()
+
+    for mapping_iter in mapping_torch:
+        mapping_iter[3] = getattr(mapping_iter[0], mapping_iter[1], None)
+        setattr(mapping_iter[0], mapping_iter[1], mapping_iter[2])
+
+
+def unpatch_torch():
+    """
+    unpatch_torch is used to unpatch optimized torch classes to original ones.
+    """
+    # TODO: unpatch_torch to support gpu-to-cpu unpatching
+
+    mapping_torch = _get_patch_map()
+
+    for mapping_iter in mapping_torch:
+        setattr(mapping_iter[0], mapping_iter[1], mapping_iter[3])
